Added new things, the log in page
//TODO
Create controllers for all the pages
connect the sql.
Design the pages
Display the database into the page

//MIGHT CHANGE
the classes and iterface.
needs somethign similar from the Book project that we had in class.

//Some changes
package instead of directory inside the java folder

I can make a database and table now, just need to conect it